* commit '8084af89d283ed015e28a5eb781726ae5e138cf0':
  nl80211: Register read_sta_data() handler for station only builds

+++ wpa_supplicant/ctrl_iface.c
@@ -4331,6 +4331,7 @@ static int wpa_supplicant_signal_poll(struct wpa_supplicant *wpa_s, char *buf,
 	return ret;
 }
 
+
 static int wpa_supplicant_pktcnt_poll(struct wpa_supplicant *wpa_s, char *buf,
 				      size_t buflen)
 {
@@ -4342,8 +4343,8 @@ static int wpa_supplicant_pktcnt_poll(struct wpa_supplicant *wpa_s, char *buf,
 		return -1;
 
 	ret = os_snprintf(buf, buflen, "TXGOOD=%lu\nTXBAD=%lu\nRXGOOD=%lu\n",
-	            sta.tx_packets, sta.tx_retry_failed, sta.rx_packets);
-	if (ret < 0 || (unsigned int) ret > buflen)
+			  sta.tx_packets, sta.tx_retry_failed, sta.rx_packets);
+	if (ret < 0 || (size_t) ret > buflen)
 		return -1;
 	return ret;
 }
